Added a helper file for FilterBarCard component.
